Bug 1370344 - Part 3: Add/refactor types of custom_properties for Properties & Values. r=#style
Add a new ExtraData enum. We carry it along with a custom property's
specified value in order to compute it, if it ends up being registered
as being able to contain URLs through Properties & Values. When the
specified value comes from a declaration, we keep track of the
associated UrlExtraData in the Specified variant. However, specified
values will also  be able to come from animations: in that case we will
eventually (when we add support for registered custom properties to
animations) be able to carry along a copy of the computed value in the
Precomputed variant, to save us from recomputation and also to save us
from having to re-resolve the URL.

The properties_and_values module is addded by a later patch in this
series, so we omit the Precomputed variant for now.

Differential Revision: https://phabricator.services.mozilla.com/D30860