Bug 1370344 - Part 4: Add a properties_and_values module for CSS Properties & Values. r=#style
Typed custom properties are exactly the same as regular custom
properties except when it comes to inheritance, computed values, and
animation, so this module is intended to be as independent from
custom_properties as possible (they do use types from each other,
though, where it would be clumsy to do otherwise).

This commit doesn't hook up the module to anything: that's handled by
later patches in this series.

Differential Revision: https://phabricator.services.mozilla.com/D30861