Bug 1370344 - Part 10: Return initial values of typed properties in CSSOM. r=#style
The initial value for unregistered custom properties is the (invalid)
empty token stream. We also store the (possibly non-empty) initial value
of registered custom properties this way. This means that during
variable substitution and access through CSSOM, we must return the
initial value that was registered, if any, rather than the stored empty
value. We already implement the former; now implement the latter.

Differential Revision: https://phabricator.services.mozilla.com/D30867